+++ tests/Core/Validation/Validators/XmlValidatorTest.php
@@ -47,16 +47,6 @@ class XmlValidatorTest extends TestCase
         return $validator;
     }
 
-    /**
-     * @depends testConstructor
-     */
-    public function testSetSpec(XmlValidator $validator): void
-    {
-        $spec = $this->createStub(Spec::class);
-        $validator->setSpec($spec);
-        self::assertSame($spec, $validator->getSpec());
-    }
-
     public function testValidateString(): void
     {
         $validator = $this->createPartialMock(XmlValidator::class, ['validateDom']);